+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Add `evidenceSubmittable` to `Dispute`
 * Add `merchantTokenIdentifier`, `sourceCardLast4` to `ApplePayCard` and `ApplePayDetails`
 * Add `threeDSecureAuthenticationId` to `CustomerRequest` and `PaymentMethodRequest`
+* Fix failing specs by passing `transaction_source` as estimated
 
 ## 3.24.0
 * Allow vaulting of Android Pay network tokens via Customer.create/update